playground/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
regex==2021.10.23
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.7.2
python-dateutil==2.8.2
python-dotenv==0.19.2
defusedxml==0.7.1
numpy==1.18.5
pillow==9.0.1
